Region e Squad
15/10

+++ src/main/java/ControllerManipularMain.java
@@ -27,16 +27,32 @@ public class ControllerManipularMain extends ControllerUtil {
     @FXML
     private void enviarDados() throws IOException, GeneralSecurityException {
         GoogleDrive gd = new GoogleDrive();
-
+        int segure = 0;
 
         //deleta os velhos primeiros (só deleta caso o arquivo existir no drive <- ARRUMAR ISSO)
-        gd.deleteFile("region");
-        gd.deleteFile("squad");
+        try {
+            gd.getFile("application/json", "region.json");
+        } catch (IOException e) {
+            segure = 1;
+        }
+        if (segure == 0) {
+            gd.deleteFile("region");
+            segure=0;
+        }
+
+        try {
+            gd.getFile("application/json", "squad.json");
+        } catch (IOException e) {
+            segure = 2;
+        }
+        if (segure == 0) {
+            gd.deleteFile("squad");
+        }
 
         //substitui os novos
         //primeiro envia o squad dps o region
-        gd.uploadFile(System.getProperty("user.dir") + File.separator + "data" + File.separator + "squadJ", "squad");
-        gd.uploadFile(System.getProperty("user.dir") + File.separator + "data" + File.separator + "regionJ", "region");
+        gd.uploadFile("region.json", System.getProperty("user.dir") + File.separator + "data" + File.separator, "application/json");
+        gd.uploadFile("squad.json", System.getProperty("user.dir") + File.separator + "data" + File.separator, "application/json");
 
     }
